New i18n/l10n function: _().
This function first tries to get a translation for the passed string in
the current bundle. If no string is found, then it tries to fallback to
the en-US string. Finally, if neither of the methods succeeded, it
returns the given string but only starting since the last dot.

So for instance if there are no translations for
_('euskalbar.menu.General'), the function will return 'General' as a
very last fallback string.

svn path=/trunk/; revision=286